本次提交内容：分类坐标发送

+++ new_project/new_peoject/code/communication.h
@@ -48,11 +48,11 @@ typedef struct{
 }Card;
 
 extern int now_distance_x;
-extern unsigned long now_distance_y;
+extern unsigned int now_distance_y;
 extern Card card_position[100];
 extern float Card_angle;//捕获到卡片时的角度
 extern float delta_card_angle;
-extern int center_distance;//卡片的直线距离  
+extern float center_distance;//卡片的直线距离  
 
 extern int card_word_ready;//卡片世界坐标是否解算完毕
 extern unsigned int card_count;
